[MM-21216] Improve visiblity of IsValidId validation function
