+++ model/utils.go
@@ -19,7 +19,6 @@ import (
 	"strconv"
 	"strings"
 	"time"
-	"unicode"
 
 	goi18n "github.com/mattermost/go-i18n/i18n"
 	"github.com/pborman/uuid"
@@ -539,20 +538,6 @@ func IsValidTrueOrFalseString(value string) bool {
 func IsValidNumberString(value string) bool {
 	if _, err := strconv.Atoi(value); err != nil {
 		return false
-	}
-
-	return true
-}
-
-func IsValidId(value string) bool {
-	if len(value) != 26 {
-		return false
-	}
-
-	for _, r := range value {
-		if !unicode.IsLetter(r) && !unicode.IsNumber(r) {
-			return false
-		}
 	}
 
 	return true
