+++ model/utils_test.go
@@ -466,39 +466,6 @@ func TestIsValidAlphaNumHyphenUnderscore(t *testing.T) {
 	}
 }
 
-func TestIsValidId(t *testing.T) {
-	cases := []struct {
-		Input  string
-		Result bool
-	}{
-		{
-			Input:  NewId(),
-			Result: true,
-		},
-		{
-			Input:  "",
-			Result: false,
-		},
-		{
-			Input:  "junk",
-			Result: false,
-		},
-		{
-			Input:  "qwertyuiop1234567890asdfg{",
-			Result: false,
-		},
-		{
-			Input:  NewId() + "}",
-			Result: false,
-		},
-	}
-
-	for _, tc := range cases {
-		actual := IsValidId(tc.Input)
-		require.Equalf(t, actual, tc.Result, "case: %v\tshould returned: %#v", tc, tc.Result)
-	}
-}
-
 func TestNowhereNil(t *testing.T) {
 	t.Parallel()
